Initial commit.
Much of this code is abased on a previous iteration of this project.
However the previous iteration was not designed to be made public or
support easy adoption. Substantial changes have been made in this
version in order to support the ultimate goal of making this code
public, versatile, and easy to adopt.